Exact and Inexact Methods for Graph Similarity in Structural Pattern Recognition
Graphs are widely employed in many application fields, such as biology,
chemistry, social networks, databases and so on. Where data are composed of
objects connected to each other and the properties of a datum depends not
only on the objects it contains, but also on how they are interconnected. Analyzing these data often requires to measure the similarity between two graphs.
Unfortunately, due to its combinatorial nature, this is a NP-Complete problem
generally addressed by using graph matching algorithms. Nevertheless, under
certain conditions and using heuristics which depend on the kind of graphs the
algorithm is facing, the solution may be computed in a polynomial time.
In this Thesis we have explored two approaches to compute the similarity
between graphs. The former is based on the exact graph matching. We have designed, VF3, an algorithm aimed to search for pattern structures within graphs.
While, the former is an inexact graph matching method which aims to compute
the Graph Edit Distance (GED) as Quadratic Assignment Problem (QAP). To
this intent, we have designed an algorithm to compute the GED by optimizing the Integer Projection Fixed Point (IPFP), a generic algorithm to solve the
QAP.